+++ src/components/widgets/Join/Join.tsx
@@ -1,11 +1,9 @@
 import { FC } from "react";
 
-import classes from "./Join.module.css";
 import useJoin from "./hooks/useJoin";
+import { JoinProps } from "@/components/widgets/Join/types/JoinProps";
 
-interface JoinProps {
-    event_id: string;
-}
+import classes from "./Join.module.css";
 
 export const Join: FC<JoinProps> = ({ event_id }) => {
     const { joinedUsers } = useJoin({ event_id });
