+++ src/components/widgets/Join/hooks/useJoin.ts
@@ -1,63 +1,13 @@
 import { useMutation, useQuery } from "@apollo/client";
-import gql from "graphql-tag";
 
-interface useJoinProps {
-    event_id: string;
-    user_id?: string | null;
-}
+import { JOIN_MUTATION } from "@/app/api/graphql/mutations/join";
+import { JOINED_BY_USER, JOINED_BY_USERS } from "@/app/api/graphql/queries/joined";
 
-type IJoined = {
-    _id: string;
-    event_id: string;
-    user_id: string;
-    isJoined: boolean;
-    createdAt?: string;
-    updatedAt?: string;
-};
+import { JoinProps } from "@/components/widgets/Join/types/JoinProps";
+import { IJoined, IGetJoined } from "@/components/widgets/Join/types/IJoined";
 
-export interface IGetJoinedQuery {
-    joinedByUsers: IJoined[];
-}
-
-const JOIN_EVENT_MUTATION = gql`
-    mutation Mutation($eventId: ID!, $userId: ID!) {
-        joinEvent(event_id: $eventId, user_id: $userId) {
-            _id
-            event_id
-            user_id
-            isJoined
-            createdAt
-            updatedAt
-        }
-    }
-`;
-
-const JOINED_BY_USERS = gql`
-    query GetJoinedByUsers($eventId: ID) {
-        joinedByUsers(event_id: $eventId) {
-            _id
-            event_id
-            user_id
-            isJoined
-        }
-    }
-`;
-
-const JOINED_BY_USER = gql`
-    query JoinedByUser($eventId: ID!, $userId: ID!) {
-        joinedByUser(event_id: $eventId, user_id: $userId) {
-            _id
-            event_id
-            user_id
-            isJoined
-            createdAt
-            updatedAt
-        }
-    }
-`;
-
-const useJoin = ({ event_id, user_id }: useJoinProps) => {
-    const { data, refetch: refetchJoinedUsers } = useQuery<IGetJoinedQuery | null>(JOINED_BY_USERS, {
+const useJoin = ({ event_id, user_id }: JoinProps) => {
+    const { data, refetch: refetchJoinedUsers } = useQuery<IGetJoined | null>(JOINED_BY_USERS, {
         variables: { eventId: event_id },
     });
 
@@ -71,7 +21,7 @@ const useJoin = ({ event_id, user_id }: useJoinProps) => {
     const joinedUsers = data?.joinedByUsers.length;
     const isJoined = joinedByUser?.joinedByUser?.isJoined;
 
-    const [joinEventMutation] = useMutation(JOIN_EVENT_MUTATION);
+    const [joinEventMutation] = useMutation(JOIN_MUTATION);
 
     const handleJoin = async () => {
         try {
