+++ src/components/widgets/Join/types/JoinProps.ts
@@ -1,0 +1,4 @@
+export interface JoinProps {
+    event_id: string;
+    user_id?: string | null;
+}
